refactor: crate dir
