+++ frontend/rust-lib/flowy-core/src/config.rs
@@ -9,8 +9,7 @@ use flowy_server_pub::af_cloud_config::AFCloudConfiguration;
 use flowy_user::services::entities::URL_SAFE_ENGINE;
 use lib_infra::file_util::copy_dir_recursive;
 use lib_infra::util::OperatingSystem;
-
-use crate::integrate::log::create_log_filter;
+use crate::log_filter::create_log_filter;
 
 #[derive(Clone)]
 pub struct AppFlowyCoreConfig {
