+++ frontend/rust-lib/flowy-core/src/deps_resolve/mod.rs
@@ -8,10 +8,12 @@ pub use user_deps::*;
 
 mod collab_deps;
 mod document_deps;
-mod folder_deps;
 
 mod chat_deps;
 mod database_deps;
 pub mod file_storage_deps;
 mod search_deps;
 mod user_deps;
+mod folder_deps;
+mod cloud_service_impl;
+pub(crate) mod reminder_deps;
