--- frontend/rust-lib/flowy-core/src/integrate/mod.rs
@@ -1,5 +1,0 @@
-pub(crate) mod collab_interact;
-pub mod log;
-pub(crate) mod server;
-mod trait_impls;
-pub(crate) mod user;
